Renamed and move editor script for the component system
